geothermal-rag/requirements.txt
# Core dependencies
gradio>=4.0.0
pyyaml>=6.0

# LLM and embeddings
ollama>=0.1.0
chromadb>=0.4.0

# PDF processing
PyMuPDF>=1.23.0

# NLP and text processing
spacy>=3.7.0
sentence-transformers>=2.2.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
